json: Initial commit (buggy)
So, the idea here was to map JSON to Lua's native data types. There's
a big problem with that, because:

  1) what does the Lua table {} represent? Empty JSON array, or an
     empty JSON object?

  2) try for k,v in pairs({1,nil,3}) do print(k,v) end, or
         for k,v in pairs({foo=nil}) do print(k,v) end, and tell me how
     to map that to JSON

It's probably better to use jansson's json_t* as a uservalue, and associate
it with a metatable implementing the __index and __newindex metamethods.
Not as clean as native data structures, but we're at least able to not mangle
structures. __gc should be pretty easy with jansson's reference counting.