freebsd/vm: Remove the N flag from yans.log file
newsyslog rotates log files by creating a new file and rename(2)-ing
it to the actual file. syslogd keeps an open file descriptor to
the file that was replaced by rename(2). Only when receiving SIGHUP
is the file descriptors closed and re-opened. The N flag causes
newsyslog to not send SIGHUP to syslogd.

I did not understand this mechanism, and when reading the man page
for newsyslog.conf:

     N       indicates that there is no process which needs to be
             signaled when this log file is rotated.

I assumed I could/should set this flag.