x509: Work-around OpenSSL bug
In multiple versions of OpenSSL, BIO_NOCLOSE is either ignored (causing
a use-after-free in caller code) or causes a memleak. This is a known
bug[1].

This commit copies the data with strdup(3) (assumes it does not contain
a \0-byte, which the rest of the SAN code assumes anyway) and free's it
later with free(3) instead.

[1] https://github.com/openssl/openssl/issues/8375